fix: reorganize files richText
